Split code into several files
